# BlazorWebSite/Swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Database API",
    "version": "1.0"
  },
  "paths": {
    "/api/BEV/DeleteEVSpecs/{id}": {
      "delete": {
        "tags": [
          "BEV"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/BEV/PatchEVSpecs/{id}": {
      "patch": {
        "tags": [
          "BEV"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "nullable": true
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EVSpecs"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/EVSpecs"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/EVSpecs"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/BEV/PostEVSpecs": {
      "post": {
        "tags": [
          "BEV"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EVSpecs"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/EVSpecs"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/EVSpecs"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/BEV/GetSpecs/{spec}/{availableOnly}": {
      "get": {
        "tags": [
          "BEV"
        ],
        "parameters": [
          {
            "name": "spec",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "availableOnly",
            "in": "path",
            "required": true,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ChartData"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ChartData"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ChartData"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/BEV/EVSpecs": {
      "get": {
        "tags": [
          "BEV"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EVSpecs"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EVSpecs"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EVSpecs"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Contact/GetContact/{email}": {
      "get": {
        "tags": [
          "Contact"
        ],
        "parameters": [
          {
            "name": "email",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Contact"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Contact"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Contact"
                }
              }
            }
          }
        }
      }
    },
    "/api/Contact/GetContactById/{id}": {
      "get": {
        "tags": [
          "Contact"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Contact"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Contact"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Contact"
                }
              }
            }
          }
        }
      }
    },
    "/api/Contact/DeleteContact/{id}": {
      "delete": {
        "tags": [
          "Contact"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Contact/PatchContact/{id}": {
      "patch": {
        "tags": [
          "Contact"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "nullable": true
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Contact"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Contact"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Contact"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Contact/PutContact": {
      "put": {
        "tags": [
          "Contact"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Contact"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Contact"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Contact"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Contact/login": {
      "post": {
        "tags": [
          "Contact"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Contact"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Contact"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Contact"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Contact"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Contact"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Contact"
                }
              }
            }
          }
        }
      }
    },
    "/api/Contact/Contacts": {
      "get": {
        "tags": [
          "Contact"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Contact"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Contact"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Contact"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Contact/ContactsInSubdivision/{subdivision}": {
      "get": {
        "tags": [
          "Contact"
        ],
        "parameters": [
          {
            "name": "subdivision",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Contact"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Contact"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Contact"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "BodyStyle": {
        "enum": [
          0,
          1,
          2,
          3,
          4,
          5,
          6,
          7,
          8,
          9,
          10
        ],
        "type": "integer",
        "format": "int32"
      },
      "DriveType": {
        "enum": [
          0,
          1,
          2,
          3,
          4
        ],
        "type": "integer",
        "format": "int32"
      },
      "MotorPowerUnits": {
        "enum": [
          0,
          1
        ],
        "type": "integer",
        "format": "int32"
      },
      "ChargingConnector": {
        "enum": [
          0,
          1,
          2,
          3
        ],
        "type": "integer",
        "format": "int32"
      },
      "EVSpecs": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "dateUpdated": {
            "type": "string",
            "format": "date-time"
          },
          "modelYear": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "manufacturer": {
            "type": "string",
            "nullable": true
          },
          "available": {
            "type": "boolean"
          },
          "model": {
            "type": "string",
            "nullable": true
          },
          "bodyStyle": {
            "$ref": "#/components/schemas/BodyStyle"
          },
          "price": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "federalTaxCredit": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "driveTrain": {
            "$ref": "#/components/schemas/DriveType"
          },
          "combinedRange": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "cityRange": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "hiwayRange": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "motorPowerKw": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "motorPowerUnits": {
            "$ref": "#/components/schemas/MotorPowerUnits"
          },
          "torque": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "batteryCapacity": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "chargingConnector": {
            "$ref": "#/components/schemas/ChargingConnector"
          },
          "weight": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "zeroTo60mph": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "zeroTo100kph": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "maxChargePower": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "minutesTo80PercentCharge": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "safetyRating": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "notes": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ChartData": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "y": {
            "type": "number",
            "format": "double",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "InterestLevel": {
        "enum": [
          0,
          1,
          2,
          3
        ],
        "type": "integer",
        "format": "int32"
      },
      "ContactStatus": {
        "enum": [
          0,
          1,
          2,
          3
        ],
        "type": "integer",
        "format": "int32"
      },
      "Contact": {
        "type": "object",
        "properties": {
          "subscriber": {
            "type": "boolean",
            "nullable": true
          },
          "id": {
            "type": "string",
            "nullable": true
          },
          "password": {
            "type": "string",
            "nullable": true
          },
          "hashedPassword": {
            "type": "string",
            "nullable": true
          },
          "dateUpdated": {
            "type": "string",
            "format": "date-time"
          },
          "evInterest": {
            "$ref": "#/components/schemas/InterestLevel"
          },
          "lawnEquipmentInterest": {
            "$ref": "#/components/schemas/InterestLevel"
          },
          "status": {
            "$ref": "#/components/schemas/ContactStatus"
          },
          "firstName": {
            "type": "string",
            "nullable": true
          },
          "lastName": {
            "type": "string",
            "nullable": true
          },
          "subdivision": {
            "type": "string",
            "nullable": true
          },
          "streetNumber": {
            "type": "integer",
            "format": "int32"
          },
          "street": {
            "type": "string",
            "nullable": true
          },
          "city": {
            "type": "string",
            "nullable": true
          },
          "state": {
            "type": "string",
            "nullable": true
          },
          "zipCode": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "phone": {
            "type": "string",
            "nullable": true
          },
          "notes": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}